Moved card specific css to own files
