+++ src/app/shared/components/Cards/Entities/Loadout/LoadoutCard.css
@@ -1,0 +1,13 @@
+.loadout-card {
+	width: 100%;
+	min-height: 175px;
+}
+
+.loadout-card:hover {
+	transform: scale(1.015);
+	transition: transform ease-in 0.15s;
+}
+
+.loadout-card:not(:first-child) {
+	margin-top: 20px;
+}
