+++ src/app/shared/components/Cards/Entities/Loadout/LoadoutCard.js
@@ -5,7 +5,9 @@ import Card from '@material-ui/core/Card'
 import CardHeader from '@material-ui/core/CardHeader'
 import CardContent from '@material-ui/core/CardContent'
 
-import CardDeleteButton from '../CardDeleteButton'
+import CardDeleteButton from '../../CardDeleteButton'
+
+import './LoadoutCard.css'
 
 export default class LoadoutCard extends React.Component {
 	render() {
